fix: improvement of modular structure
