+++ js/main.js
@@ -1,4 +1,3 @@
-import './constants.js';
 import './functions/get-random-integer.js';
 import './functions/get-random-array-element.js';
 import './functions/unique-number-generator.js';
